refactor #2: 모집 금액 필드 수정
- amount를 price로 변경하여 단가 의미 명확화
- 금액 연산이 없어 BigDecimal 대신 Long 사용

+++ src/main/java/com/cu2mber/recruitmentservice/dto/response/InternalRecruitmentSummaryResponse.java
@@ -7,8 +7,6 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
 
-import java.math.BigDecimal;
-
 @Getter
 @ToString
 @NoArgsConstructor
@@ -25,10 +23,10 @@ public class InternalRecruitmentSummaryResponse {
 
     private String recruitmentTitle;
 
-    private BigDecimal price;
+    private Long price;
 
     @QueryProjection
-    public InternalRecruitmentSummaryResponse(Long recruitmentNo, StatusType recruitmentStatus, Long memberLocalNo, Long eventNo, String recruitmentTitle, BigDecimal price) {
+    public InternalRecruitmentSummaryResponse(Long recruitmentNo, StatusType recruitmentStatus, Long memberLocalNo, Long eventNo, String recruitmentTitle, Long price) {
         this.recruitmentNo = recruitmentNo;
         this.recruitmentStatus = recruitmentStatus;
         this.memberLocalNo = memberLocalNo;
